requirements.txt
fastapi~=0.112.2
uvicorn
pytube~=15.0.0
langchain~=0.2.14
langchain_community
openai~=1.42.0
git+https://github.com/openai/whisper.git
sentence-transformers
pinecone-client
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
python-dotenv~=1.0.1
youtube-transcript-api
pydantic~=2.8.2
starlette~=0.38.2
langdetect~=1.0.9
requests~=2.32.3